uploading files
I've edited the index file from how it is in the tutorial. it is now instead for a social media platform for musicians. I also added pages for the submit button and terms and conditions so they dont direct to freecodecamp. I plan at editing the css file in the future